LayoutTests/editing/pasteboard/paste-noscript-xhtml-expected.txt

ALERT: hello
ALERT: hello

This test copies the content of an iframe and pastes it
in an editable area in an xhtml document and verifies that no script, handlers or javascript urls are copied.


Copied content:
| "
"
| <button>
|   id="button1"
|   onclick="sayHello()"
|   ondblclick="sayHello()"
|   style="width: 100px;"
|   "<#selection-anchor>Hello"
| <a>
|   href="http://www.bing.com/search?q=cnn"
|   id="anchor1"
|   "CNN"
| <a>
|   href="javascript:sayHello()"
|   id="anchor2"
|   "Hello"
| "
"
| <script>
|   style="display: block; position: fixed; top: 4000px"
|   "alert('hello');"
| "
"
| <script>
|   src="../editing.js"
|   style="display: block; position: fixed; top: 4000px"
| "
"
| <iframe>
|   id="iframe1"
|   src="javascript:alert('hello')"
|   style="width: 200px; height: 100px; background-color:#cee;"
| "
"
| <form>
|   action="javascript:sayHello()"
|   formaction="javascript:sayHello()"
|   id="form1"
|   style="width: 200px; height: 150px; background-color:#cee;"
|   "This is a form<#selection-focus>"
| "

"

FRAME 0:
| <head>
| <body>

Pasted content:
| <button>
|   id="button1"
|   style="width: 100px; "
|   "Hello"
| <a>
|   href="http://www.bing.com/search?q=cnn"
|   id="anchor1"
|   "CNN"
| <a>
|   href=""
|   id="anchor2"
|   "Hello"
| <iframe>
|   id="iframe1"
|   src=""
|   style="width: 200px; height: 100px; background-color: rgb(204, 238, 238); "
| <form>
|   action=""
|   formaction=""
|   id="form1"
|   style="width: 200px; height: 150px; background-color: rgb(204, 238, 238); "
|   "This is a form<#selection-caret>"

FRAME 0:
| <head>
| <body>
